feat: only trigger UTXO scanning when a new block event is received
Previously the UTXO scanner task was triggered on an interval. This PR updates this process to rather be triggered only when an existing task is not running and new Block event is received from the wallet Base Node monitoring service. It is only when this event is received when we know there will be new block data to scan.

A new event is added to the Base Node Monitoring Service that only fires when a new block is detected to be used in the above functionality

The Utxo Scanner is also refactored into separate files. To help in the review I have highlighted the new code below

An unrelated bug is also fixed with handling of the wallet public address in the Console Wallet that broke the wallet.
