+++ base_layer/wallet/src/base_node_service/monitor.rs
@@ -88,7 +88,7 @@ where
                 },
                 Err(e @ BaseNodeMonitorError::RpcFailed(_)) => {
                     warn!(target: LOG_TARGET, "Connectivity failure to base node: {}", e);
-                    self.map_state(move |_| BaseNodeState {
+                    self.update_state(BaseNodeState {
                         chain_metadata: None,
                         is_synced: None,
                         updated: None,
@@ -134,7 +134,7 @@ where
             let tip_info = match interrupt(base_node_watch.changed(), client.get_tip_info()).await {
                 Some(tip_info) => tip_info?,
                 None => {
-                    self.map_state(|_| Default::default()).await;
+                    self.update_state(Default::default()).await;
                     continue;
                 },
             };
@@ -165,7 +165,8 @@ where
 
             let is_synced = tip_info.is_synced;
             let height_of_longest_chain = chain_metadata.height_of_longest_chain();
-            self.map_state(move |_| BaseNodeState {
+
+            self.update_state(BaseNodeState {
                 chain_metadata: Some(chain_metadata),
                 is_synced: Some(is_synced),
                 updated: Some(Utc::now().naive_utc()),
@@ -184,7 +185,7 @@ where
 
             let delay = time::sleep(self.interval.saturating_sub(latency));
             if interrupt(base_node_watch.changed(), delay).await.is_none() {
-                self.map_state(|_| Default::default()).await;
+                self.update_state(Default::default()).await;
             }
         }
 
@@ -193,14 +194,23 @@ where
         Ok(())
     }
 
-    async fn map_state<F>(&self, transform: F)
-    where F: FnOnce(&BaseNodeState) -> BaseNodeState {
-        let new_state = {
-            let mut lock = self.state.write().await;
-            let new_state = transform(&*lock);
-            *lock = new_state.clone();
-            new_state
+    async fn update_state(&self, new_state: BaseNodeState) {
+        let mut lock = self.state.write().await;
+        let (new_block_detected, height) = match (new_state.chain_metadata.clone(), (*lock).chain_metadata.clone()) {
+            (Some(new_metadata), Some(old_metadata)) => (
+                new_metadata.height_of_longest_chain() != old_metadata.height_of_longest_chain(),
+                new_metadata.height_of_longest_chain(),
+            ),
+            (Some(new_metadata), _) => (true, new_metadata.height_of_longest_chain()),
+            (None, _) => (false, 0),
         };
+
+        if new_block_detected {
+            self.publish_event(BaseNodeEvent::NewBlockDetected(height));
+        }
+
+        *lock = new_state.clone();
+
         self.publish_event(BaseNodeEvent::BaseNodeStateChanged(new_state));
     }
