+++ base_layer/wallet/src/utxo_scanner_service/mod.rs
@@ -1,38 +1,18 @@
-// Copyright 2021. The Tari Project
-//
-// Redistribution and use in source and binary forms, with or without modification, are permitted provided that the
-// following conditions are met:
-//
-// 1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following
-// disclaimer.
-//
-// 2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the
-// following disclaimer in the documentation and/or other materials provided with the distribution.
-//
-// 3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote
-// products derived from this software without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
-// INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-// DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-// SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
-// WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
-// USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
 use crate::{
+    base_node_service::handle::BaseNodeServiceHandle,
     connectivity_service::{WalletConnectivityHandle, WalletConnectivityInterface},
     output_manager_service::handle::OutputManagerHandle,
     storage::database::{WalletBackend, WalletDatabase},
     transaction_service::handle::TransactionServiceHandle,
     utxo_scanner_service::{
         handle::UtxoScannerHandle,
-        utxo_scanning::{UtxoScannerMode, UtxoScannerService},
+        service::UtxoScannerService,
+        uxto_scanner_service_builder::UtxoScannerMode,
     },
 };
 use futures::future;
 use log::*;
-use std::{sync::Arc, time::Duration};
+use std::sync::Arc;
 use tari_comms::{connectivity::ConnectivityRequester, NodeIdentity};
 use tari_core::transactions::CryptoFactories;
 use tari_service_framework::{async_trait, ServiceInitializationError, ServiceInitializer, ServiceInitializerContext};
@@ -40,12 +20,15 @@ use tokio::sync::broadcast;
 
 pub mod error;
 pub mod handle;
-pub mod utxo_scanning;
+pub mod service;
+mod utxo_scanner_task;
+mod uxto_scanner_service_builder;
+
+pub use utxo_scanner_task::RECOVERY_KEY;
 
 const LOG_TARGET: &str = "wallet::utxo_scanner_service::initializer";
 
 pub struct UtxoScannerServiceInitializer<T> {
-    interval: Duration,
     backend: Option<WalletDatabase<T>>,
     factories: CryptoFactories,
     node_identity: Arc<NodeIdentity>,
@@ -54,14 +37,8 @@ pub struct UtxoScannerServiceInitializer<T> {
 impl<T> UtxoScannerServiceInitializer<T>
 where T: WalletBackend + 'static
 {
-    pub fn new(
-        interval: Duration,
-        backend: WalletDatabase<T>,
-        factories: CryptoFactories,
-        node_identity: Arc<NodeIdentity>,
-    ) -> Self {
+    pub fn new(backend: WalletDatabase<T>, factories: CryptoFactories, node_identity: Arc<NodeIdentity>) -> Self {
         Self {
-            interval,
             backend: Some(backend),
             factories,
             node_identity,
@@ -87,7 +64,6 @@ where T: WalletBackend + 'static
             .take()
             .expect("Cannot start Utxo scanner service without setting a storage backend");
         let factories = self.factories.clone();
-        let interval = self.interval;
         let node_identity = self.node_identity.clone();
 
         context.spawn_when_ready(move |handles| async move {
@@ -95,11 +71,11 @@ where T: WalletBackend + 'static
             let output_manager_service = handles.expect_handle::<OutputManagerHandle>();
             let comms_connectivity = handles.expect_handle::<ConnectivityRequester>();
             let wallet_connectivity = handles.expect_handle::<WalletConnectivityHandle>();
+            let base_node_service_handle = handles.expect_handle::<BaseNodeServiceHandle>();
 
             let scanning_service = UtxoScannerService::<T>::builder()
                 .with_peers(vec![])
                 .with_retry_limit(2)
-                .with_scanning_interval(interval)
                 .with_mode(UtxoScannerMode::Scanning)
                 .build_with_resources(
                     backend,
@@ -111,6 +87,7 @@ where T: WalletBackend + 'static
                     factories,
                     handles.get_shutdown_signal(),
                     event_sender,
+                    base_node_service_handle,
                 )
                 .run();
