+++ base_layer/wallet/src/wallet.rs
@@ -80,7 +80,7 @@ use crate::{
         TransactionServiceInitializer,
     },
     types::KeyDigest,
-    utxo_scanner_service::{handle::UtxoScannerHandle, UtxoScannerServiceInitializer},
+    utxo_scanner_service::{handle::UtxoScannerHandle, UtxoScannerServiceInitializer, RECOVERY_KEY},
 };
 use tari_common_types::transaction::TxId;
 use tari_key_manager::cipher_seed::CipherSeed;
@@ -188,7 +188,6 @@ where
             ))
             .add_initializer(WalletConnectivityInitializer::new(config.base_node_service_config))
             .add_initializer(UtxoScannerServiceInitializer::new(
-                config.scan_for_utxo_interval,
                 wallet_database.clone(),
                 factories.clone(),
                 node_identity.clone(),
@@ -498,7 +497,6 @@ where
     /// Utility function to find out if there is data in the database indicating that there is an incomplete recovery
     /// process in progress
     pub async fn is_recovery_in_progress(&self) -> Result<bool, WalletError> {
-        use crate::utxo_scanner_service::utxo_scanning::RECOVERY_KEY;
         Ok(self.db.get_client_key_value(RECOVERY_KEY.to_string()).await?.is_some())
     }
 }
