+++ base_layer/wallet_ffi/src/lib.rs
@@ -153,7 +153,7 @@ use tari_wallet::{
             models::{CompletedTransaction, InboundTransaction, OutboundTransaction},
         },
     },
-    utxo_scanner_service::utxo_scanning::{UtxoScannerService, RECOVERY_KEY},
+    utxo_scanner_service::{service::UtxoScannerService, RECOVERY_KEY},
     Wallet,
     WalletConfig,
     WalletSqlite,
@@ -3212,7 +3212,6 @@ pub unsafe extern "C" fn wallet_create(
         }),
         None,
         Network::Weatherwax.into(),
-        None,
         None,
         None,
         None,
