+++ common/src/configuration/global.rs
@@ -101,7 +101,6 @@ pub struct GlobalConfig {
     pub fetch_utxos_timeout: Duration,
     pub service_request_timeout: Duration,
     pub base_node_query_timeout: Duration,
-    pub scan_for_utxo_interval: Duration,
     pub saf_expiry_duration: Duration,
     pub transaction_broadcast_monitoring_timeout: Duration,
     pub transaction_chain_monitoring_timeout: Duration,
@@ -445,11 +444,6 @@ fn convert_node_config(
         cfg.get_int(key)
             .map_err(|e| ConfigurationError::new(key, &e.to_string()))? as u64,
     );
-    let key = "wallet.scan_for_utxo_interval";
-    let scan_for_utxo_interval = Duration::from_secs(
-        cfg.get_int(key)
-            .map_err(|e| ConfigurationError::new(key, &e.to_string()))? as u64,
-    );
 
     let key = "wallet.saf_expiry_duration";
     let saf_expiry_duration = Duration::from_secs(optional(cfg.get_int(key))?.unwrap_or(10800) as u64);
@@ -755,7 +749,6 @@ fn convert_node_config(
         fetch_utxos_timeout,
         service_request_timeout,
         base_node_query_timeout,
-        scan_for_utxo_interval,
         saf_expiry_duration,
         transaction_broadcast_monitoring_timeout,
         transaction_chain_monitoring_timeout,
