+++ common/src/configuration/utils.rs
@@ -118,7 +118,6 @@ pub fn default_config(bootstrap: &ConfigBootstrap) -> Config {
     cfg.set_default("wallet.base_node_service_refresh_interval", 5).unwrap();
     cfg.set_default("wallet.base_node_service_request_max_age", 60).unwrap();
     cfg.set_default("wallet.balance_enquiry_cooldown_period", 1).unwrap();
-    cfg.set_default("wallet.scan_for_utxo_interval", 60 * 60 * 12).unwrap();
     cfg.set_default("wallet.transaction_broadcast_monitoring_timeout", 60)
         .unwrap();
     cfg.set_default("wallet.transaction_chain_monitoring_timeout", 60)
